editing notifications fix

+++ db/db_manager.py
@@ -18,8 +18,8 @@ def fetchall(columns: List[str]) -> List[Tuple]:
     notifications = Notification.select().dicts().execute()
     return notifications
 
-def clean_done_tasks() -> None:
-    query = Notification.delete().where(Notification.is_done == True)
+def clean_done_tasks(user_id: int) -> None:
+    query = Notification.delete().where(Notification.user_id == user_id).where(Notification.is_done == True)
     query.execute()
 
 def delete_by_id(id: int) -> Tuple:
@@ -76,6 +76,36 @@ def edit_title_with_id(id: int, data: str):
     notification = Notification(text = data)
     notification.notification_id = id
     notification.save()
+
+def edit_frequency_with_id(id: int, frequncy: str):
+    notification = Notification(frequency = frequncy)
+    notification.notification_id = id
+    notification.save()
+
+def add_attachment_with_id(id: int, attachment: str):
+    notification = Notification.get(Notification.notification_id == id)
+    new_attachments = notification.attachments + attachment
+    updated_notification = Notification(attachments = new_attachments)
+    updated_notification.notification_id = id
+    updated_notification.save()
+
+def del_attachments_with_id(id: int, indices: list) -> int:
+    print(indices)
+    notification = Notification.get(Notification.notification_id == id)
+    attachments = notification.attachments.split(';')
+    attachments.pop(-1)
+    # print(*attachments)
+    if min(indices) < 0 or max(indices) > len(attachments):
+        return -1
+    new_attachments = list(att for idx, att in enumerate(attachments) if not idx in indices)
+    # print(*new_attachments)
+    new_attachments_str = ';'.join(new_attachments)
+    new_attachments_str += ';'
+    # print(new_attachments_str)
+    notification = Notification(attachments=new_attachments_str)
+    notification.notification_id = id
+    notification.save()
+    return 0
 
 def find_by_date(date: str) -> List:
     notifications = Notification.select().where(Notification.time == date).dicts().execute()
